# Dashboard/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Sharecycle API",
    "description": "API documentation for the Sharecycle backend application.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/payment/orders": {
      "post": {
        "tags": [
          "checkout-controller"
        ],
        "summary": "Step 1: Create a PayPal Order and Get Approval URL",
        "description": "Initiates the payment process for a bicycle rental.\n\nThis endpoint performs two main actions:\n1.  Creates a corresponding **Order entity** within our system with a `PENDING` status.\n2.  Creates a secure **Order with PayPal** and returns a unique `approvalUrl`.\n\nThe frontend should then redirect the user to this `approvalUrl` to approve the payment on PayPal's website.\n",
        "operationId": "pay",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PayResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/payment/orders/{paypalOrderID}/authorize": {
      "post": {
        "tags": [
          "checkout-controller"
        ],
        "summary": "Step 2: Authorize Payment After User Approval",
        "description": "This endpoint should be called by the frontend immediately after the user approves the payment on PayPal and is redirected back to the application.\n\nIt finalizes the pre-approved payment:\n1.  It calls PayPal's API to officially **authorize** the transaction using the `paypalOrderID`.\n2.  It retrieves the unique `authorizationId` from PayPal.\n3.  The **Payment Status** in our database is updated from `INITIALIZED` to `AUTHORIZED`, and the `authorizationId` is stored for future use (i.e., for capturing the funds or voiding authorization).\n\nThis action places a hold on the user's funds, which is valid for 29 days.\n",
        "operationId": "authorizeOrder",
        "parameters": [
          {
            "name": "paypalOrderID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/fcm/token": {
      "post": {
        "tags": [
          "FCM"
        ],
        "summary": "Register FCM token",
        "description": "Register or update FCM device token for push notifications",
        "operationId": "registerToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FCMTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "FCM"
        ],
        "summary": "Remove FCM token",
        "description": "Remove FCM token when user logs out",
        "operationId": "removeToken",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/fcm/test": {
      "post": {
        "tags": [
          "FCM"
        ],
        "operationId": "test",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "iphoneNumber": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "idCardNumber": {
            "type": "string"
          },
          "idCardImageUrl": {
            "type": "string"
          }
        }
      },
...
      "OrderRequestDTO": {
        "type": "object",
        "properties": {
          "bicycleId": {
            "type": "integer",
            "format": "int64"
          },
          "dateFrom": {
            "type": "string"
          },
          "dateTo": {
            "type": "string"
          }
        }
      },
      "OrderResponseDTO": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer",
            "format": "int64"
          },
          "bicycleId": {
            "type": "integer",
            "format": "int64"
          },
          "bikeName": {
            "type": "string"
          },
          "bikeImageUrl": {
            "type": "string"
          },
          "pricePerDay": {
            "type": "number",
            "format": "double"
          },
          "renter": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "owner": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "dateFrom": {
            "type": "string"
          },
          "dateTo": {
            "type": "string"
          },
          "rentalDays": {
            "type": "integer",
            "format": "int64"
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "orderDate": {
            "type": "string",
            "format": "date"
          },
          "orderStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED",
              "CANCELED"
            ]
          }
        }
      },
...
      "PayResponseDTO": {
        "type": "object",
        "properties": {
          "approvalUrl": {
            "type": "string"
          },
          "paypalOrderId": {
            "type": "string"
          }
        }
      },
      "FCMTokenRequest": {
        "required": [
          "fcmToken"
        ],
        "type": "object",
        "properties": {
          "fcmToken": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}